refactor: HomeScreenを編集
