+++ lib/ui/screens/home_screen.dart
@@ -1,0 +1,109 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:flutter_training/models/weather_condition.dart';
+import 'package:flutter_training/ui/extensions/weather_condition_ext.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+class HomeScreen extends StatefulWidget {
+  const HomeScreen({super.key});
+
+  @override
+  State<HomeScreen> createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen> {
+  final YumemiWeather _yumemiWeather = YumemiWeather();
+  WeatherCondition? _weatherCondition;
+
+  Future<void> _getWeather() async {
+    final condition = _yumemiWeather.fetchSimpleWeather();
+    setState(() {
+      _weatherCondition = WeatherCondition.from(condition);
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: FractionallySizedBox(
+          widthFactor: 0.5,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              AspectRatio(
+                aspectRatio: 1,
+                child: _weatherCondition == null
+                    ? const Placeholder()
+                    : SvgPicture.asset(_weatherCondition!.svgAsset),
+              ),
+              const SizedBox(height: 16),
+              const _TemperatureRow(),
+              const SizedBox(height: 80),
+              _ButtonRow(getWeather: _getWeather),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _ButtonRow extends StatelessWidget {
+  const _ButtonRow({required VoidCallback getWeather})
+      : _onReloadButtonPressed = getWeather;
+  final VoidCallback _onReloadButtonPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Expanded(
+          child: TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: const Text('Close'),
+          ),
+        ),
+        Expanded(
+          child: TextButton(
+            onPressed: _onReloadButtonPressed,
+            child: const Text('Reload'),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class _TemperatureRow extends StatelessWidget {
+  const _TemperatureRow();
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Expanded(
+          child: Text(
+            '**℃',
+            textAlign: TextAlign.center,
+            style: Theme.of(context).textTheme.labelLarge?.copyWith(
+                  color: Theme.of(context).colorScheme.primary,
+                ),
+          ),
+        ),
+        Expanded(
+          child: Text(
+            '**℃',
+            textAlign: TextAlign.center,
+            style: Theme.of(context).textTheme.labelLarge?.copyWith(
+                  color: Theme.of(context).colorScheme.error,
+                ),
+          ),
+        ),
+      ],
+    );
+  }
+}
